BREAKING CHANGE: split configuration in multiple files
