+++ lua/plugins/dressing.lua
@@ -1,0 +1,6 @@
+return {
+  {
+    "stevearc/dressing.nvim",
+    opts = {},
+  },
+}
