+++ lua/plugins/mason-lspconfig.lua
@@ -1,0 +1,10 @@
+return {
+  {
+    "williamboman/mason-lspconfig.nvim",
+    config = function()
+      require("mason-lspconfig").setup({
+        ensure_installed = { "lua_ls", "jdtls", "phpactor" }, -- Add your servers here
+      })
+    end,
+  },
+}
