+++ lua/plugins/nui.line.lua
@@ -1,0 +1,5 @@
+return {
+  {
+  "MunifTanjim/nui.nvim",
+  lazy = true,
+} }
