+++ lua/plugins/nvim-dap-virtual-text.lua
@@ -1,0 +1,6 @@
+return {
+  {
+    "theHamsta/nvim-dap-virtual-text",
+    opts = {},
+  },
+}
