+++ lua/plugins/nvim-spectre.lua
@@ -1,0 +1,11 @@
+return {
+  {
+    "nvim-pack/nvim-spectre",
+    cmd = "Spectre",
+    opts = { open_cmd = "noswapfile vnew" },
+    -- stylua: ignore
+    keys = {
+      { "<leader>Sr", function() require("spectre").open() end, desc = "Replace in files (Spectre)" },
+    },
+  },
+}
