+++ lua/plugins/nvim-web-devicons.lua
@@ -1,0 +1,6 @@
+return {
+  {
+    "nvim-tree/nvim-web-devicons",
+    lazy = true,
+  },
+}
