+++ lua/plugins/overseer.lua
@@ -1,0 +1,8 @@
+return {
+  {
+    "stevearc/overseer.nvim",
+    config = function()
+      require("overseer").setup()
+    end,
+  },
+}
