+++ lua/plugins/supermaven-nvim.lua
@@ -1,0 +1,15 @@
+return {
+  {
+    "supermaven-inc/supermaven-nvim",
+    lazy = true,
+    config = function()
+      require("supermaven-nvim").setup({
+        keymaps = {
+          accept_suggestion = "<C-j>",
+          clear_suggestion = "<C-]>",
+          accept_word = "<C-.>",
+        },
+      })
+    end,
+  },
+}
