+++ lua/plugins/which-key.lua
@@ -1,0 +1,25 @@
+return {
+  {
+    "folke/which-key.nvim",
+    event = "VeryLazy",
+    opts = function(_, opts)
+      local wk = require("which-key")
+      wk.add({
+        { "<leader>sn", group = "+noice" },
+        { "<leader>s", group = "+search" },
+        { "<leader>g", group = "+git" },
+        { "<leader>q", group = "+quit/session" },
+        { "<leader>t", group = "+terminal" },
+      })
+    end,
+    keys = {
+      {
+        "<leader>?",
+        function()
+          require("which-key").show({ global = false })
+        end,
+        desc = "Buffer Local Keymaps (which-key)",
+      },
+    },
+  },
+}
